backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
mysql-connector-python
passlib[bcrypt]
python-jose 